+++ jest.config.cjs
@@ -1,0 +1,4 @@
+module.exports = {
+	preset: 'ts-jest',
+	testEnvironment: 'node',
+};
